proxy/requirements.txt
# Meshtastic Proxy GUI Requirements

# Qt GUI Framework
PyQt6>=6.4.0

# Meshtastic Protocol
meshtastic>=2.2.0

# For Gemini AI integration (optional)
# Using the new Google Generative AI SDK v2 (google-genai)
google-genai>=0.2.0

# Telegram Bot integration (optional)
python-telegram-bot>=20.0

# Async networking
asyncio-mqtt>=0.16.1

# Build tool
pyinstaller>=5.13.0
